Migrating parts of the old `Node` class to new interface
The following parts have been moved:

 * Methods interacting with attributes
 * Methods interacting with the label and description
 * Methods related to links
 * Methods related to storing a node

Additionally, actual business logic is removed from the legacy workflow
database models, because they were referencing the node code that is to
be migrated, but the model classes are only kept such that a migration
to drop the tables is not yet necessary. The API to run legacy workflows
had already been dropped in a previous commit and is no longer supported.
