+++ aiida/common/lang.py
@@ -118,6 +118,18 @@ class EmptyContextManager(object):  # pylint: disable=too-few-public-methods,use
         pass
 
 
-def type_check(what, of_type):
+def type_check(what, of_type, msg=None):
+    """
+    Check if object 'what' is of type 'of_type'.
+    If not, raise a TypeError.
+
+    :param what: the object to check
+    :param of_type: the type (or tuple of types) to compare to
+    :param msg: if specified, allows to customize the message that
+        is passed within the TypeError exception
+    """
+
     if not isinstance(what, of_type):
-        raise TypeError("Got object of type '{}', expecting '{}'".format(type(what), of_type))
+        if msg is None:
+            "Got object of type '{}', expecting '{}'".format(type(what), of_type)
+        raise TypeError(msg)
