+++ aiida/orm/implementation/django/oldnode.py
@@ -27,50 +27,29 @@ from aiida.orm.implementation.general.node import AbstractNode, _HASH_EXTRA_KEY
 
 class Node(AbstractNode):
 
-    @classmethod
-    def get_subclass_from_uuid(cls, uuid):
-        from .convert import get_backend_entity
-        from aiida.backends.djsite.db.models import DbNode
-        try:
-            node = get_backend_entity(DbNode.objects.get(uuid=uuid), None)
-        except ObjectDoesNotExist:
-            raise NotExistent("No entry with UUID={} found".format(uuid))
-        if not isinstance(node, cls):
-            raise NotExistent("UUID={} is not an instance of {}".format(uuid, cls.__name__))
-        return node
+    # @classmethod
+    # def get_subclass_from_uuid(cls, uuid):
+    #     from .convert import get_backend_entity
+    #     from aiida.backends.djsite.db.models import DbNode
+    #     try:
+    #         node = get_backend_entity(DbNode.objects.get(uuid=uuid), None)
+    #     except ObjectDoesNotExist:
+    #         raise NotExistent("No entry with UUID={} found".format(uuid))
+    #     if not isinstance(node, cls):
+    #         raise NotExistent("UUID={} is not an instance of {}".format(uuid, cls.__name__))
+    #     return node
 
-    @classmethod
-    def get_subclass_from_pk(cls, pk):
-        from .convert import get_backend_entity
-        from aiida.backends.djsite.db.models import DbNode
-        try:
-            node = get_backend_entity(DbNode.objects.get(pk=pk), None)
-        except ObjectDoesNotExist:
-            raise NotExistent("No entry with pk= {} found".format(pk))
-        if not isinstance(node, cls):
-            raise NotExistent("pk= {} is not an instance of {}".format(pk, cls.__name__))
-        return node
-
-    @classmethod
-    def query(cls, *args, **kwargs):
-        from aiida.backends.djsite.db.models import DbNode
-        if cls._plugin_type_string:
-            if not cls._plugin_type_string.endswith('.'):
-                raise InternalError("The plugin type string does not " "finish with a dot??")
-
-            # If it is 'calculation.Calculation.', we want to filter
-            # for things that start with 'calculation.' and so on
-            plug_type = cls._plugin_type_string
-
-            # Remove the implementation.django or sqla part.
-            if plug_type.startswith('implementation.'):
-                plug_type = '.'.join(plug_type.split('.')[2:])
-            pre, sep, _ = plug_type[:-1].rpartition('.')
-            superclass_string = "".join([pre, sep])
-            return DbNode.aiidaobjects.filter(*args, type__startswith=superclass_string, **kwargs)
-        else:
-            # Base Node class, with empty string
-            return DbNode.aiidaobjects.filter(*args, **kwargs)
+    # @classmethod
+    # def get_subclass_from_pk(cls, pk):
+    #     from .convert import get_backend_entity
+    #     from aiida.backends.djsite.db.models import DbNode
+    #     try:
+    #         node = get_backend_entity(DbNode.objects.get(pk=pk), None)
+    #     except ObjectDoesNotExist:
+    #         raise NotExistent("No entry with pk= {} found".format(pk))
+    #     if not isinstance(node, cls):
+    #         raise NotExistent("pk= {} is not an instance of {}".format(pk, cls.__name__))
+    #     return node
 
     def __init__(self, **kwargs):
         from aiida.backends.djsite.db.models import DbNode
@@ -131,345 +110,3 @@ class Node(AbstractNode):
             # Automatically set all *other* attributes, if possible, otherwise
             # stop
             self._set_with_defaults(**kwargs)
-
-    @classmethod
-    def query(cls, *args, **kwargs):
-        from aiida.backends.djsite.db.models import DbNode
-        if cls._plugin_type_string:
-            if not cls._plugin_type_string.endswith('.'):
-                raise InternalError("The plugin type string does not " "finish with a dot??")
-
-            # If it is 'calculation.Calculation.', we want to filter
-            # for things that start with 'calculation.' and so on
-            plug_type = cls._plugin_type_string
-
-            # Remove the implementation.django or sqla part.
-            if plug_type.startswith('implementation.'):
-                plug_type = '.'.join(plug_type.split('.')[2:])
-            pre, sep, _ = plug_type[:-1].rpartition('.')
-            superclass_string = "".join([pre, sep])
-            return DbNode.aiidaobjects.filter(*args, type__startswith=superclass_string, **kwargs)
-        else:
-            # Base Node class, with empty string
-            return DbNode.aiidaobjects.filter(*args, **kwargs)
-
-
-    @property
-    def nodeversion(self):
-        return self._dbnode.nodeversion
-
-    @property
-    def ctime(self):
-        return self._dbnode.ctime
-
-    @property
-    def mtime(self):
-        return self._dbnode.mtime
-
-
-    def _db_label_field(self):
-        return self._dbnode.label
-
-    def _update_db_label_field(self, field_value):
-        self._dbnode.label = field_value
-        if self.is_stored:
-            with transaction.atomic():
-                self._dbnode.save()
-                self._increment_version_number_db()
-
-    def _get_db_description_field(self):
-        return self._dbnode.description
-
-    def _update_db_description_field(self, field_value):
-        self._dbnode.description = field_value
-        if self.is_stored:
-            with transaction.atomic():
-                self._dbnode.save()
-                self._increment_version_number_db()
-
-    def _add_dblink_from(self, src, link_type, label):
-        from aiida.orm.querybuilder import QueryBuilder
-        if not isinstance(src, Node):
-            raise ValueError("src must be a Node instance")
-        if self.uuid == src.uuid:
-            raise ValueError("Cannot link to itself")
-
-        if not src.is_stored:
-            raise ModificationNotAllowed("Cannot call the internal _add_dblink_from if the "
-                                         "source node is not stored")
-
-        if not self.is_stored:
-            raise ModificationNotAllowed("Cannot call the internal _add_dblink_from if the "
-                                         "destination node is not stored")
-
-        if link_type is LinkType.CREATE or link_type is LinkType.INPUT_CALC or link_type is LinkType.INPUT_WORK:
-            # Check for cycles. This works if the transitive closure is enabled; if it
-            # isn't, this test will never fail, but then having a circular link is not
-            # meaningful but does not pose a huge threat
-            #
-            # I am linking src->self; a loop would be created if a DbPath exists already
-            # in the TC table from self to src
-            if QueryBuilder().append(
-                    Node, filters={
-                        'id': self.pk
-                    }, tag='parent').append(
-                        Node, filters={
-                            'id': src.pk
-                        }, tag='child', with_ancestors='parent').count() > 0:
-                raise ValueError("The link you are attempting to create would generate a loop")
-
-        self._do_create_link(src, label, link_type)
-
-    def _do_create_link(self, src, label, link_type):
-        from aiida.backends.djsite.db.models import DbLink
-
-        sid = None
-        try:
-            # transactions are needed here for Postgresql:
-            # https://docs.djangoproject.com/en/1.5/topics/db/transactions/#handling-exceptions-within-postgresql-transactions
-            sid = transaction.savepoint()
-            DbLink.objects.create(input=src._dbnode, output=self._dbnode, label=label, type=link_type.value)
-            transaction.savepoint_commit(sid)
-        except IntegrityError as exc:
-            transaction.savepoint_rollback(sid)
-            raise UniquenessError("There is already a link with the same " "name (raw message was {})" "".format(exc))
-
-    def _get_db_input_links(self, link_type):
-        from aiida.orm.convert import get_orm_entity
-        from aiida.backends.djsite.db.models import DbLink
-
-        link_filter = {'output': self._dbnode}
-        if link_type is not None:
-            link_filter['type'] = link_type.value
-        return [(i.label, get_orm_entity(i.input)) for i in DbLink.objects.filter(**link_filter).distinct()]
-
-    def _get_db_output_links(self, link_type):
-        from aiida.orm.convert import get_orm_entity
-        from aiida.backends.djsite.db.models import DbLink
-
-        link_filter = {'input': self._dbnode}
-        if link_type is not None:
-            link_filter['type'] = link_type.value
-        return ((i.label, get_orm_entity(i.output)) for i in DbLink.objects.filter(**link_filter).distinct())
-
-
-
-    def _set_db_attr(self, key, value):
-        """
-        Set the value directly in the DB, without checking if it is stored, or
-        using the cache.
-
-        DO NOT USE DIRECTLY.
-
-        :param str key: key name
-        :param value: its value
-        """
-        from aiida.backends.djsite.db.models import DbAttribute
-
-        DbAttribute.set_value_for_node(self._dbnode, key, value)
-        self._increment_version_number_db()
-
-    def _del_db_attr(self, key):
-        from aiida.backends.djsite.db.models import DbAttribute
-        if not DbAttribute.has_key(self._dbnode, key):
-            raise AttributeError("DbAttribute {} does not exist".format(key))
-        DbAttribute.del_value_for_node(self._dbnode, key)
-        self._increment_version_number_db()
-
-    def _get_db_attr(self, key):
-        from aiida.backends.djsite.db.models import DbAttribute
-        return DbAttribute.get_value_for_node(dbnode=self._dbnode, key=key)
-
-    def _db_iterattrs(self):
-        from aiida.backends.djsite.db.models import DbAttribute
-
-        all_attrs = DbAttribute.get_all_values_for_node(self._dbnode)
-        for attr in all_attrs:
-            yield (attr, all_attrs[attr])
-
-    def _db_attrs(self):
-        # Note: I "duplicate" the code from iterattrs and reimplement it
-        # here, rather than
-        # calling iterattrs from here, because iterattrs is slow on each call
-        # since it has to call .getvalue(). To improve!
-        from aiida.backends.djsite.db.models import DbAttribute
-        attrlist = DbAttribute.list_all_node_elements(self._dbnode)
-        for attr in attrlist:
-            yield attr.key
-
-    def _increment_version_number_db(self):
-        from aiida.backends.djsite.db.models import DbNode
-        # I increment the node number using a filter
-        self._dbnode.nodeversion = F('nodeversion') + 1
-        self._dbnode.save()
-
-        # This reload internally the node of self._dbnode
-        # Note: I have to reload the object (to have the right values in memory,
-        # otherwise I only get the Django Field F object as a result!
-        self._dbnode = DbNode.objects.get(pk=self._dbnode.pk)
-
-    @property
-    def uuid(self):
-        return six.text_type(self._dbnode.uuid)
-
-    @property
-    def id(self):
-        return self._dbnode.id
-
-    @property
-    def process_type(self):
-        return self._dbnode.process_type
-
-    @property
-    def dbnode(self):
-        # I also update the internal _dbnode variable, if it was saved
-        # from aiida.backends.djsite.db.models import DbNode
-        #        if self.is_stored:
-        #            self._dbnode = DbNode.objects.get(pk=self._dbnode.pk)
-        return self._dbnode
-
-    def _db_store_all(self, with_transaction=True, use_cache=None):
-        """
-        Store the node, together with all input links, if cached, and also the
-        linked nodes, if they were not stored yet.
-
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-        """
-        from django.db import transaction
-        from aiida.common.lang import EmptyContextManager
-
-        if with_transaction:
-            context_man = transaction.atomic()
-        else:
-            context_man = EmptyContextManager()
-
-        with context_man:
-            # Always without transaction: either it is the context_man here,
-            # or it is managed outside
-            self._store_input_nodes()
-            self.store(with_transaction=False, use_cache=use_cache)
-            self._store_cached_input_links(with_transaction=False)
-
-        return self
-
-
-    def _store_cached_input_links(self, with_transaction=True):
-        """
-        Store all input links that are in the local cache, transferring them
-        to the DB.
-
-        :note: This can be called only if all parents are already stored.
-
-        :note: Links are stored only after the input nodes are stored. Moreover,
-            link storage is done in a transaction, and if one of the links
-            cannot be stored, an exception is raised and *all* links will remain
-            in the cache.
-
-        :note: This function can be called only after the node is stored.
-           After that, it can be called multiple times, and nothing will be
-           executed if no links are still in the cache.
-
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-        """
-        from django.db import transaction
-        from aiida.common.lang import EmptyContextManager
-
-        if with_transaction:
-            context_man = transaction.atomic()
-        else:
-            context_man = EmptyContextManager()
-
-        if not self.is_stored:
-            raise ModificationNotAllowed('cannot store cached incoming links for unstored Node<{}>'.format(self.pk))
-
-        with context_man:
-            # This raises if there is an unstored node.
-            self._check_are_parents_stored()
-
-            # I have to store only those links where the source is already stored
-            for link_triple in self._incoming_cache:
-                self._add_dblink_from(*link_triple)
-
-            # If everything went smoothly, clear the entries from the cache.
-            # I do it here because I delete them all at once if no error
-            # occurred; otherwise, links will not be stored and I
-            # should not delete them from the cache (but then an exception
-            # would have been raised, and the following lines are not executed)
-            self._incoming_cache = list()
-
-    def _db_store(self, with_transaction=True):
-        """
-        Store a new node in the DB, also saving its repository directory
-        and attributes.
-
-        After being called attributes cannot be
-        changed anymore! Instead, extras can be changed only AFTER calling
-        this store() function.
-
-        :note: After successful storage, those links that are in the cache, and
-            for which also the parent node is already stored, will be
-            automatically stored. The others will remain unstored.
-
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-
-        :param bool use_cache: Whether I attempt to find an equal node in the DB.
-        """
-        # TODO: This needs to be generalized, allowing for flexible methods
-        # for storing data and its attributes.
-        from django.db import transaction
-        from aiida.common.lang import EmptyContextManager
-        from aiida.backends.djsite.db.models import DbAttribute
-
-        if with_transaction:
-            context_man = transaction.atomic()
-        else:
-            context_man = EmptyContextManager()
-
-        # I save the corresponding django entry
-        # I set the folder
-        # NOTE: I first store the files, then only if this is successful,
-        # I store the DB entry. In this way,
-        # I assume that if a node exists in the DB, its folder is in place.
-        # On the other hand, periodically the user might need to run some
-        # bookkeeping utility to check for lone folders.
-        self._repository_folder.replace_with_folder(self._get_temp_folder().abspath, move=True, overwrite=True)
-
-        # I do the transaction only during storage on DB to avoid timeout
-        # problems, especially with SQLite
-        try:
-            with context_man:
-                # Save the row
-                self._dbnode.save()
-                # Save its attributes 'manually' without incrementing
-                # the version for each add.
-                DbAttribute.reset_values_for_node(self._dbnode, attributes=self._attrs_cache, with_transaction=False)
-                # This should not be used anymore: I delete it to
-                # possibly free memory
-                del self._attrs_cache
-
-                self._temp_folder = None
-                self._to_be_stored = False
-
-                # Here, I store those links that were in the cache and
-                # that are between stored nodes.
-                self._store_cached_input_links()
-
-        # This is one of the few cases where it is ok to do a 'global'
-        # except, also because I am re-raising the exception
-        except:
-            # I put back the files in the sandbox folder since the
-            # transaction did not succeed
-            self._get_temp_folder().replace_with_folder(self._repository_folder.abspath, move=True, overwrite=True)
-            raise
-
-        from aiida.backends.djsite.db.models import DbExtra
-        # I store the hash without cleaning and without incrementing the nodeversion number
-        DbExtra.set_value_for_node(self._dbnode, _HASH_EXTRA_KEY, self.get_hash())
-
-        return self
