+++ aiida/orm/nodes.py
@@ -30,8 +30,9 @@ from aiida.common.lang import combomethod, classproperty, type_check
 from aiida.common.escaping import sql_string_match
 from aiida.manage import get_manager
 from aiida.manage.caching import get_use_cache
-from aiida.orm.utils.node import AbstractNodeMeta, clean_value, _NO_DEFAULT
+from aiida.orm.utils.node import AbstractNodeMeta
 from aiida.orm.utils.managers import NodeInputManager, NodeOutputManager
+from aiida.orm.implementation.nodes import _NO_DEFAULT
 from . import comments
 from . import convert
 from . import entities
@@ -67,13 +68,6 @@ class Node(entities.Entity):
     # This will be set by the metaclass call
     _logger = None
 
-    # Name to be used for the Repository section
-    _section_name = 'node'
-
-    # The name of the subfolder in which to put the files/directories
-    # added with add_path
-    _path_subfolder_name = 'path'
-
     # A list of tuples, saying which attributes cannot be set at the same time
     # See documentation in the set() method.
     _set_incompatibilities = []
@@ -88,12 +82,7 @@ class Node(entities.Entity):
     # Flag that determines whether the class can be cached.
     _cacheable = True
 
-    # Flag that says if the node is storable or not.
-    # By default, bare nodes (and also ProcessNodes) are not storable,
-    # all subclasses (WorkflowNode, CalculationNode, Data and their subclasses)
-    # are storable. This flag is checked in store()
-    _storable = False
-    _unstorable_message = 'only Data, WorkflowNode, CalculationNode or their subclasses can be stored'
+    # region VariousMethods
 
     def get_desc(self):
         """
@@ -231,23 +220,6 @@ class Node(entities.Entity):
         # Can be redefined in the subclasses
         self._init_internal_params()
 
-        self._attrs_cache = {}
-
-        # A cache of incoming links represented as a list of LinkTriples instances
-        self._incoming_cache = list()
-
-        self._temp_folder = None
-        self._repo_folder = RepositoryFolder(section=self._section_name, uuid=self.uuid)
-
-    def __del__(self):
-        """
-        Called only upon real object destruction from memory
-        I just try to remove junk, whenever possible; do not trust
-        too much this function!
-        """
-        if getattr(self, '_temp_folder', None) is not None:
-            self._temp_folder.erase()
-
     def __repr__(self):
         return '<{}: {}>'.format(self.__class__.__name__, str(self))
 
@@ -279,114 +251,96 @@ class Node(entities.Entity):
         """
         return self._backend_entity.mtime
 
-    def has_cached_links(self):
-        """
-        Return whether there are unstored incoming links in the cache.
+    # TODO: setting initial and internal parameters still to check
+    # def _init_internal_params(self):
+    #     """
+    #     Set the default values for this class; this method is automatically called by the init.
 
-        :return: boolean, True when there are links in the incoming cache, False otherwise
-        """
-        return bool(self._incoming_cache)
+    #     :note: if you inherit this function, ALWAYS remember to
+    #       call super()._init_internal_params() as the first thing
+    #       in your inherited function.
+    #     """
+    #     pass
 
-    def _init_internal_params(self):
-        """
-        Set the default values for this class; this method is automatically called by the init.
+    # @property
+    # def _set_defaults(self):
+    #     """
+    #     Default values to set in the __init__, if no value is explicitly provided for the given key.
+    #     It is a dictionary, with k=v; if the key k is not provided to the __init__, and a value is present here,
+    #     this is set.
+    #     """
+    #     return {}
 
-        :note: if you inherit this function, ALWAYS remember to
-          call super()._init_internal_params() as the first thing
-          in your inherited function.
-        """
-        pass
+    # def _set_with_defaults(self, **kwargs):
+    #     """
+    #     Calls the set() method, but also adds the class-defined default
+    #     values (defined in the self._set_defaults attribute),
+    #     if they are not provided by the user.
 
-    @property
-    def _set_defaults(self):
-        """
-        Default values to set in the __init__, if no value is explicitly provided for the given key.
-        It is a dictionary, with k=v; if the key k is not provided to the __init__, and a value is present here,
-        this is set.
-        """
-        return {}
+    #     :note: for the default values, also allow to define 'hidden' methods,
+    #         meaning that if a default value has a key "_state", it will not call
+    #         the function "set__state" but rather "_set_state".
+    #         This is not allowed, instead, for the standard set() method.
+    #     """
+    #     self._set_internal(arguments=self._set_defaults, allow_hidden=True)
 
-    @classmethod
-    def query(cls, *_args, **_kwargs):
-        """
-        Query for nodes of this type
+    #     # Pass everything to 'set'
+    #     self.set(**kwargs)
 
-        .. deprecated:: 1.0.0b1
-            Use :meth:`aiida.orm.Node.objects.query` instead.
-        """
-        warnings.warn("use aidia.orm.Node.objects.query instead", DeprecationWarning)
-        raise NotImplementedError("This method is no longer available, use Node.objects.query()")
+    # def set(self, **kwargs):
+    #     """
+    #     For each k=v pair passed as kwargs, call the corresponding
+    #     set_k(v) method (e.g., calling self.set(property=5, mass=2) will
+    #     call self.set_property(5) and self.set_mass(2).
+    #     Useful especially in the __init__.
 
-    def _set_with_defaults(self, **kwargs):
-        """
-        Calls the set() method, but also adds the class-defined default
-        values (defined in the self._set_defaults attribute),
-        if they are not provided by the user.
+    #     :note: it uses the _set_incompatibilities list of the class to check
+    #         that we are not setting methods that cannot be set at the same time.
+    #         _set_incompatibilities must be a list of tuples, and each tuple
+    #         specifies the elements that cannot be set at the same time.
+    #         For instance, if _set_incompatibilities = [('property', 'mass')],
+    #         then the call self.set(property=5, mass=2) will raise a ValueError.
+    #         If a tuple has more than two values, it raises ValueError if *all*
+    #         keys are provided at the same time, but it does not give any error
+    #         if at least one of the keys is not present.
 
-        :note: for the default values, also allow to define 'hidden' methods,
-            meaning that if a default value has a key "_state", it will not call
-            the function "set__state" but rather "_set_state".
-            This is not allowed, instead, for the standard set() method.
-        """
-        self._set_internal(arguments=self._set_defaults, allow_hidden=True)
+    #     :note: If one element of _set_incompatibilities is a tuple with only
+    #         one element, this element will not be settable using this function
+    #         (and in particular,
 
-        # Pass everything to 'set'
-        self.set(**kwargs)
+    #     :raise ValueError: if the corresponding set_k method does not exist
+    #         in self, or if the methods cannot be set at the same time.
+    #     """
+    #     self._set_internal(arguments=kwargs, allow_hidden=False)
 
-    def set(self, **kwargs):
-        """
-        For each k=v pair passed as kwargs, call the corresponding
-        set_k(v) method (e.g., calling self.set(property=5, mass=2) will
-        call self.set_property(5) and self.set_mass(2).
-        Useful especially in the __init__.
+    # def _set_internal(self, arguments, allow_hidden=False):
+    #     """
+    #     Works as self.set(), but takes a dictionary as the 'arguments' variable,
+    #     instead of reading it from the ``kwargs``; moreover, it allows to specify
+    #     allow_hidden to True. In this case, if a a key starts with and
+    #     underscore, as for instance ``_state``, it will not call
+    #     the function ``set__state`` but rather ``_set_state``.
+    #     """
+    #     for incomp in self._set_incompatibilities:
+    #         if all(k in arguments.keys() for k in incomp):
+    #             if len(incomp) == 1:
+    #                 raise ValueError("Cannot set {} directly when creating "
+    #                                  "the node or using the .set() method; "
+    #                                  "use the specific method instead.".format(incomp[0]))
+    #             else:
+    #                 raise ValueError("Cannot set {} at the same time".format(" and ".join(incomp)))
 
-        :note: it uses the _set_incompatibilities list of the class to check
-            that we are not setting methods that cannot be set at the same time.
-            _set_incompatibilities must be a list of tuples, and each tuple
-            specifies the elements that cannot be set at the same time.
-            For instance, if _set_incompatibilities = [('property', 'mass')],
-            then the call self.set(property=5, mass=2) will raise a ValueError.
-            If a tuple has more than two values, it raises ValueError if *all*
-            keys are provided at the same time, but it does not give any error
-            if at least one of the keys is not present.
-
-        :note: If one element of _set_incompatibilities is a tuple with only
-            one element, this element will not be settable using this function
-            (and in particular,
-
-        :raise ValueError: if the corresponding set_k method does not exist
-            in self, or if the methods cannot be set at the same time.
-        """
-        self._set_internal(arguments=kwargs, allow_hidden=False)
-
-    def _set_internal(self, arguments, allow_hidden=False):
-        """
-        Works as self.set(), but takes a dictionary as the 'arguments' variable,
-        instead of reading it from the ``kwargs``; moreover, it allows to specify
-        allow_hidden to True. In this case, if a a key starts with and
-        underscore, as for instance ``_state``, it will not call
-        the function ``set__state`` but rather ``_set_state``.
-        """
-        for incomp in self._set_incompatibilities:
-            if all(k in arguments.keys() for k in incomp):
-                if len(incomp) == 1:
-                    raise ValueError("Cannot set {} directly when creating "
-                                     "the node or using the .set() method; "
-                                     "use the specific method instead.".format(incomp[0]))
-                else:
-                    raise ValueError("Cannot set {} at the same time".format(" and ".join(incomp)))
-
-        for key, value in arguments.items():
-            try:
-                if allow_hidden and key.startswith("_"):
-                    method = getattr(self, '_set_{}'.format(key[1:]))
-                else:
-                    method = getattr(self, 'set_{}'.format(key))
-            except AttributeError:
-                raise ValueError("Unable to set '{0}', no set_{0} method " "found".format(key))
-            if not isinstance(method, typing.Callable):
-                raise ValueError("Unable to set '{0}', set_{0} is not " "callable!".format(key))
-            method(value)
+    #     for key, value in arguments.items():
+    #         try:
+    #             if allow_hidden and key.startswith("_"):
+    #                 method = getattr(self, '_set_{}'.format(key[1:]))
+    #             else:
+    #                 method = getattr(self, 'set_{}'.format(key))
+    #         except AttributeError:
+    #             raise ValueError("Unable to set '{0}', no set_{0} method " "found".format(key))
+    #         if not isinstance(method, typing.Callable):
+    #             raise ValueError("Unable to set '{0}', set_{0} is not " "callable!".format(key))
+    #         method(value)
 
     @property
     def type(self):
@@ -467,306 +421,6 @@ class Node(entities.Entity):
         """
         return users.User.from_backend_entity(self.backend_entity.get_user())
 
-    def validate_incoming(self, source, link_type, link_label):
-        """
-        Validate adding a link of the given type from a given node to ourself.
-
-        This function will first validate the types of the inputs, followed by the node and link types and validate
-        whether in principle a link of that type between the nodes of these types is allowed.the
-
-        Subsequently, the validity of the "degree" of the proposed link is validated, which means validating the
-        number of links of the given type from the given node type is allowed.
-
-        :param source: the node from which the link is coming
-        :param link_type: the link type
-        :param link_label: the link label
-        :raise TypeError: if `source` is not a Node instance or `link_type` is not a `LinkType` enum
-        :raise ValueError: if the proposed link is invalid
-        """
-        type_check(link_type, LinkType)
-        type_check(source, Node)
-
-        from aiida.orm.utils.links import validate_link
-        validate_link(source, self, link_type, link_label)
-
-    def validate_outgoing(self, target, link_type, link_label):  # pylint: disable=unused-argument
-        """
-        Validate adding a link of the given type from ourself to a given node.
-
-        The validity of the triple (source, link, target) should be validated in the `validate_incoming` call.
-        This method will be called afterwards and can be overriden by subclasses to add additional checks that are
-        specific to that subclass.
-
-        :param target: the node to which the link is going
-        :param link_type: the link type
-        :param link_label: the link label
-        :raise TypeError: if `target` is not a Node instance or `link_type` is not a `LinkType` enum
-        :raise ValueError: if the proposed link is invalid
-        """
-        type_check(link_type, LinkType)
-        type_check(target, Node)
-
-    def add_incoming(self, source, link_type, link_label):
-        """
-        Add a link of the given type from a given node to ourself.
-
-        :param source: the node from which the link is coming
-        :param link_type: the link type
-        :param link_label: the link label
-        :return: True if the proposed link is allowed, False otherwise
-        :raise TypeError: if `source` is not a Node instance or `link_type` is not a `LinkType` enum
-        :raise ValueError: if the proposed link is invalid
-        """
-        self.validate_incoming(source, link_type, link_label)
-        source.validate_outgoing(self, link_type, link_label)
-
-        if self.is_stored and source.is_stored:
-            self.backend_entity.add_link_from(source, link_type, link_label)
-        else:
-            self._add_cachelink_from(source, link_type, link_label)
-
-    def _add_cachelink_from(self, source, link_type, link_label):
-        """Add an incoming link to the cache.
-
-        .. note: the proposed link is not validated in this function, so this should not be called directly
-            but it should only be called by `Node.add_incoming`.
-
-        :param source: the node from which the link is coming
-        :param link_type: the link type
-        :param link_label: the link label
-        """
-        link_triple = links.LinkTriple(source, link_type, link_label)
-
-        if link_triple in self._incoming_cache:
-            raise exceptions.UniquenessError('the link triple {} is already present in the cache'.format(link_triple))
-
-        self._incoming_cache.append(link_triple)
-
-    def get_incoming(self, node_class=None, link_type=(), link_label_filter=None):
-        """
-        Return a list of link triples that are (directly) incoming into this node.
-
-        :param node_class: If specified, should be a class or tuple of classes, and it filters only
-            elements of that specific type (or a subclass of 'type')
-        :param link_type: If specified should be a string or tuple to get the inputs of this
-            link type, if None then returns all inputs of all link types.
-        :param link_label_filter: filters the incoming nodes by its link label.
-            Here wildcards (% and _) can be passed in link label filter as we are using "like" in QB.
-        """
-        if not isinstance(link_type, tuple):
-            link_type = (link_type,)
-
-        if self.is_stored:
-            link_triples = self.get_stored_link_triples(node_class, link_type, link_label_filter, 'incoming')
-        else:
-            link_triples = []
-
-        # Get all cached link triples
-        for link_triple in self._incoming_cache:
-
-            if link_triple in link_triples:
-                raise exceptions.InternalError('Node<{}> has both a stored and cached link triple {}'.format(
-                    self.pk, link_triple))
-
-            if not link_type or link_triple.link_type in link_type:
-                if link_label_filter is not None:
-                    if sql_string_match(string=link_triple.link_label, pattern=link_label_filter):
-                        link_triples.append(link_triple)
-                else:
-                    link_triples.append(link_triple)
-
-        return links.LinkManager(link_triples)
-
-    def get_outgoing(self, node_class=None, link_type=(), link_label_filter=None):
-        """
-        Return a list of link triples that are (directly) outgoing of this node.
-
-        :param node_class: If specified, should be a class or tuple of classes, and it filters only
-            elements of that specific type (or a subclass of 'type')
-        :param link_type: If specified should be a string or tuple to get the inputs of this
-            link type, if None then returns all outputs of all link types.
-        :param link_label_filter: filters the outgoing nodes by its link label.
-            Here wildcards (% and _) can be passed in link label filter as we are using "like" in QB.
-        """
-        link_triples = self.get_stored_link_triples(node_class, link_type, link_label_filter, 'outgoing')
-        return links.LinkManager(link_triples)
-
-    def get_stored_link_triples(self, node_class=None, link_type=(), link_label_filter=None, link_direction='incoming'):
-        """
-        Return the list of stored link triples directly incoming to or outgoing of this node.
-
-        Note this will only return link triples that are stored in the database. Anything in the cache is ignored.
-
-        :param node_class: If specified, should be a class, and it filters only elements of that (subclass of) type
-        :param link_type: Only get inputs of this link type, if empty tuple then returns all inputs of all link types.
-        :param link_label_filter: filters the incoming nodes by its link label. This should be a regex statement as
-            one would pass directly to a QuerBuilder filter statement with the 'like' operation.
-        :param link_direction: `incoming` or `outgoing` to get the incoming or outgoing links, respectively.
-        """
-        if not isinstance(link_type, tuple):
-            link_type = (link_type,)
-
-        if link_type and not all([isinstance(t, LinkType) for t in link_type]):
-            raise TypeError('link_type should be a LinkType or tuple of LinkType: got {}'.format(link_type))
-
-        node_class = node_class or Node
-        node_filters = {'id': {'==': self.id}}
-        edge_filters = {}
-
-        if link_type:
-            edge_filters['type'] = {'in': [t.value for t in link_type]}
-
-        if link_label_filter:
-            edge_filters['label'] = {'like': link_label_filter}
-
-        builder = querybuilder.QueryBuilder()
-        builder.append(Node, filters=node_filters, tag='main')
-
-        if link_direction == 'outgoing':
-            builder.append(
-                node_class,
-                with_incoming='main',
-                project=['*'],
-                edge_project=['type', 'label'],
-                edge_filters=edge_filters)
-        else:
-            builder.append(
-                node_class,
-                with_outgoing='main',
-                project=['*'],
-                edge_project=['type', 'label'],
-                edge_filters=edge_filters)
-
-        return [links.LinkTriple(entry[0], LinkType(entry[1]), entry[2]) for entry in builder.all()]
-
-    def get_inputs_dict(self, only_in_db=False, link_type=None):
-        """
-        Return a dictionary where the key is the label of the input link, and
-        the value is the input node.
-
-        :param only_in_db: If true only get stored links, not cached
-        :param link_type: Only get inputs of this link type, if None then
-                returns all inputs of all link types.
-        :return: a dictionary {label:object}
-        """
-        warnings.warn('get_inputs_dict method is deprecated, use get_incoming instead', DeprecationWarning)
-
-        return dict(self.get_inputs(also_labels=True, only_in_db=only_in_db, link_type=link_type))
-
-    def get_outputs_dict(self, link_type=None):
-        """
-        Return a dictionary where the key is the label of the output link, and
-        the value is the input node.
-        As some Nodes (Datas in particular) can have more than one output with
-        the same label, all keys have the name of the link with appended the pk
-        of the node in output.
-        The key without pk appended corresponds to the oldest node.
-
-        :return: a dictionary {linkname:object}
-        """
-        warnings.warn('get_outputs_dict method is deprecated, use get_outgoing instead', DeprecationWarning)
-
-        if link_type is not None and not isinstance(link_type, LinkType):
-            raise TypeError("link_type should be a LinkType object")
-
-        all_outputs = self.get_outputs(also_labels=True, link_type=link_type)
-
-        all_linknames = [i[0] for i in all_outputs]
-        linknames_set = list(set(all_linknames))
-
-        # prepare a new output list
-        new_outputs = {}
-        # first add the defaults
-        for irreducible_linkname in linknames_set:
-            this_elements = [i[1] for i in all_outputs if i[0] == irreducible_linkname]
-            # select the oldest element
-            last_element = sorted(this_elements, key=lambda x: x.ctime)[0]
-            # for this one add the default value
-            new_outputs[irreducible_linkname] = last_element
-
-            # now for everyone append the string with the pk
-            for i in this_elements:
-                new_outputs[irreducible_linkname + "_{}".format(i.pk)] = i
-
-        return new_outputs
-
-    def get_inputs(self, node_type=None, also_labels=False, only_in_db=False, link_type=None):
-        """
-        Return a list of nodes that enter (directly) in this node
-
-        :param node_type: If specified, should be a class, and it filters only
-            elements of that specific type (or a subclass of 'type')
-        :param also_labels: If False (default) only return a list of input nodes.
-            If True, return a list of tuples, where each tuple has the
-            following format: ('label', Node), with 'label' the link label,
-            and Node a Node instance or subclass
-        :param only_in_db: Return only the inputs that are in the database,
-            ignoring those that are in the local cache. Otherwise, return
-            all links.
-        :param link_type: Only get inputs of this link type, if None then
-            returns all inputs of all link types.
-        """
-        warnings.warn('get_inputs method is deprecated, use get_incoming instead', DeprecationWarning)
-
-        if link_type is not None and not isinstance(link_type, LinkType):
-            raise TypeError('link_type should be a LinkType object')
-
-        inputs_list = self._backend_entity._get_db_input_links(link_type=link_type)
-
-        if not only_in_db:
-            # Needed for the check
-            input_list_keys = [i[0] for i in inputs_list]
-
-            for link_triple in self._incoming_cache:
-                if link_triple.link_label in input_list_keys:
-                    raise exceptions.InternalError(
-                        "There exist a link with the same name '{}' both in the DB and in the internal "
-                        "cache for node pk= {}!".format(link_triple.link_label, self.pk))
-
-                if link_type is None or link_triple.link_type is link_type:
-                    inputs_list.append((link_triple.link_label, link_triple.node))
-
-        if node_type is None:
-            filtered_list = inputs_list
-        else:
-            filtered_list = [i for i in inputs_list if isinstance(i[1], node_type)]
-
-        if also_labels:
-            return list(filtered_list)
-
-        return [i[1] for i in filtered_list]
-
-    def get_outputs(self, node_type=None, also_labels=False, link_type=None):
-        """
-        Return a list of nodes that exit (directly) from this node
-
-        :param node_type: if specified, should be a class, and it filters only
-            elements of that specific node_type (or a subclass of 'node_type')
-        :param also_labels: if False (default) only return a list of input nodes.
-            If True, return a list of tuples, where each tuple has the
-            following format: ('label', Node), with 'label' the link label,
-            and Node a Node instance or subclass
-        :param link_type: Only return outputs connected by links of this type.
-        """
-        warnings.warn('get_outputs method is deprecated, use get_outgoing instead', DeprecationWarning)
-
-        if link_type is not None and not isinstance(link_type, LinkType):
-            raise TypeError('link_type should be a LinkType object')
-
-        outputs_list = [
-            convert.get_orm_entity(entity) for entity in self.backend_entity.get_output_links(link_type=link_type)
-        ]
-
-        if node_type is None:
-            filtered_list = outputs_list
-        else:
-            filtered_list = (i for i in outputs_list if isinstance(i[1], node_type))
-
-        if also_labels:
-            return list(filtered_list)
-
-        return [i[1] for i in filtered_list]
-
     def get_computer(self):
         """
         Get the computer associated to the node.
@@ -778,125 +432,35 @@ class Node(entities.Entity):
         """
         self._backend_entity.get_computer()
 
-    
+    # endregion
 
     # region Attributes
 
-    def iterattrs(self):
+    @property
+    def attrs_items(self):
         """
-        Iterator over the attributes, returning tuples (key, value)
-        """
-        if not self.is_stored:
-            for key, value in self._attrs_cache.items():
-                yield (key, value)
-        else:
-            for key, value in self.backend_entity.iterattrs():
-                yield key, value
+        Iterator over the Node attributes, returning tuples (key, value)
 
-    def attrs(self):
+        .. deprecated:: 1.0.0b1
+        """
+        for item in self._backend_entity.attrs_items:
+            yield item
+
+    @property
+    def attrs_keys(self):
         """
         Returns the keys of the attributes as a generator.
 
         :return: a generator of a strings
         """
-        # Note: this calls a different function _db_attrs
-        # because often it's faster not to retrieve the values from the DB
-        if not self.is_stored:
-            for key in self._attrs_cache:
-                yield key
-        else:
-            for key in self.backend_entity.attrs():
-                yield key
+        for item in self._backend_entity.attrs_keys:
+            yield item
 
     def get_attrs(self):
         """
         Return a dictionary with all attributes of this node.
         """
-        return dict(self.iterattrs())
-
-    def _set_attr(self, key, value, clean=True, stored_check=True):
-        """
-        Set a node attribute
-
-        :param key: key name
-        :param value: its value
-        :param clean: whether to clean values.
-            WARNING: when set to False, storing will throw errors
-            for any data types not recognized by the db backend
-        :param stored_check: when set to False will disable the mutability check
-        :raise ModificationNotAllowed: if node is already stored
-        :raise ValidationError: if the key is not valid, e.g. it contains the separator symbol
-        """
-        if stored_check and self.is_stored:
-            raise exceptions.ModificationNotAllowed('Cannot change the attributes of a stored node')
-
-        validate_attribute_key(key)
-
-        if not self.is_stored:
-            if clean:
-                value = clean_value(value)
-            self._attrs_cache[key] = value
-        else:
-            self.backend_entity.set_attr(key, clean_value(value))
-
-    def _append_to_attr(self, key, value, clean=True):
-        """
-        Append value to an attribute of the Node (in the DbAttribute table).
-
-        :param key: key name of "list-type" attribute If attribute doesn't exist, it is created.
-        :param value: the value to append to the list
-        :param clean: whether to clean the value
-            WARNING: when set to False, storing will throw errors
-            for any data types not recognized by the db backend
-        :raise ValidationError: if the key is not valid, e.g. it contains the separator symbol
-        """
-        validate_attribute_key(key)
-
-        try:
-            values = self.get_attr(key)
-        except AttributeError:
-            values = []
-
-        try:
-            if clean:
-                values.append(clean_value(value))
-            else:
-                values.append(value)
-        except AttributeError:
-            raise AttributeError("Use _set_attr only on attributes containing lists")
-
-        self._set_attr(key, values, clean=False)
-
-    def _del_attr(self, key, stored_check=True):
-        """
-        Delete an attribute.
-
-        :param key: attribute to delete.
-        :param stored_check: when set to False will disable the mutability check
-        :raise AttributeError: if key does not exist.
-        :raise ModificationNotAllowed: if node is already stored
-        """
-        if stored_check and self.is_stored:
-            raise exceptions.ModificationNotAllowed('Cannot change the attributes of a stored node')
-
-        if not self.is_stored:
-            try:
-                del self._attrs_cache[key]
-            except KeyError:
-                raise AttributeError("DbAttribute {} does not exist".format(key))
-        else:
-            self.backend_entity.del_attr(key)
-
-    def _del_all_attrs(self):
-        """
-        Delete all attributes associated to this node.
-
-        :raise ModificationNotAllowed: if the Node was already stored.
-        """
-        # I have to convert the attrs in a list, because the list will change
-        # while deleting elements
-        for attr_name in list(self.attrs()):
-            self._del_attr(attr_name)
+        return dict(self.attrs_items)
 
     def get_attr(self, key, default=_NO_DEFAULT):
         """
@@ -909,18 +473,7 @@ class Node(entities.Entity):
 
         :raise AttributeError: If no attribute is found and there is no default
         """
-        try:
-            if not self.is_stored:
-                try:
-                    return self._attrs_cache[key]
-                except KeyError:
-                    raise AttributeError("DbAttribute '{}' does not exist".format(key))
-            else:
-                return self.backend_entity.get_attr(key)
-        except AttributeError:
-            if default is _NO_DEFAULT:
-                raise
-            return default
+        return self._backend_entity.get_attr(key=key, default=default)
 
     # endregion
 
@@ -955,7 +508,7 @@ class Node(entities.Entity):
             node and avoid to run multiple times the same computation on it).
         """
         self._backend_entity.set_extras(the_dict)
-        
+
     def reset_extras(self, new_extras):
         """
         Deletes existing extras and creates new ones.
@@ -978,7 +531,7 @@ class Node(entities.Entity):
         :raise ValueError: If more than two arguments are passed to get_extra
         """
         self._backend_entity.get_extra(key, *args)
-      
+
     def get_extras(self):
         """
         Get the value of extras, reading directly from the DB!
@@ -1002,21 +555,25 @@ class Node(entities.Entity):
         """
         self._backend_entity.del_extra(key)
 
-    def extras(self):
+    @property
+    def extras_items(self):
         """
-        Get the keys of the extras.
+        Iterator over the node extras, returning tuples (key, value)
 
-        :return: a list of strings
+        .. deprecated:: 1.0.0b1
         """
-        self._backend_entity.extras()
+        for item in self._backend_entity.extras_items:
+            yield item
 
-    def iterextras(self):
+    @property
+    def extras_keys(self):
         """
-        Iterator over the extras, returning tuples (key, value)
+        Returns the keys of the nodes extras as a generator.
 
-        :todo: verify that I am not creating a list internally
+        :return: a generator of a strings
         """
-        self._backend_entity.iterextras()
+        for item in self._backend_entity.extras_keys:
+            yield item
 
     # endregion
 
@@ -1029,7 +586,6 @@ class Node(entities.Entity):
         :return: the newly created comment
         """
         self._backend_entity.add_comment(content, user)
-
 
     def get_comment(self, identifier):
         """
@@ -1076,16 +632,7 @@ class Node(entities.Entity):
         """
         return self.backend_entity.uuid
 
-    @property
-    def _repository_folder(self):
-        """
-        Get the permanent repository folder.
-        Use preferentially the folder property.
-
-        :return: the permanent RepositoryFolder object
-        """
-        return self._repo_folder
-
+    # TODO: decide if we want to keep this or expose different functionality
     @property
     def folder(self):
         """
@@ -1094,99 +641,7 @@ class Node(entities.Entity):
 
         :return: the RepositoryFolder object.
         """
-        if not self.is_stored:
-            return self._get_temp_folder()
-
-        return self._repository_folder
-
-    def _get_folder_pathsubfolder(self):
-        """
-        Get the subfolder in the repository.
-
-        :return: a Folder object.
-        """
-        return self.folder.get_subfolder(self._path_subfolder_name, reset_limit=True)
-
-    def get_folder_list(self, subfolder='.'):
-        """
-        Get the the list of files/directory in the repository of the object.
-
-        :param subfolder: get the list of a subfolder
-        :return: a list of strings.
-        """
-        return self._get_folder_pathsubfolder().get_subfolder(subfolder).get_content_list()
-
-    def _get_temp_folder(self):
-        """
-        Get the folder of the Node in the temporary repository.
-
-        :return: a SandboxFolder object mapping the node in the repository.
-        """
-        # I create the temp folder only at is first usage
-        if self._temp_folder is None:
-            self._temp_folder = SandboxFolder()  # This is also created
-            # Create the 'path' subfolder in the Sandbox
-            self._get_folder_pathsubfolder().create()
-        return self._temp_folder
-
-    def remove_path(self, path):
-        """
-        Remove a file or directory from the repository directory.
-        Can be called only before storing.
-
-        :param str path: relative path to file/directory.
-        """
-        if self.is_stored:
-            raise exceptions.ModificationNotAllowed("Cannot delete a path after storing the node")
-
-        if os.path.isabs(path):
-            raise ValueError("The destination path in remove_path must be a relative path")
-        self._get_folder_pathsubfolder().remove_path(path)
-
-    def add_path(self, src_abs, dst_path):
-        """
-        Copy a file or folder from a local file inside the repository directory.
-        If there is a subpath, folders will be created.
-
-        Copy to a cache directory if the entry has not been saved yet.
-
-        :param str src_abs: the absolute path of the file to copy.
-        :param str dst_filename: the (relative) path on which to copy.
-
-        :todo: in the future, add an add_attachment() that has the same
-            meaning of a extras file. Decide also how to store. If in two
-            separate subfolders, remember to reset the limit.
-        """
-        if self.is_stored:
-            raise exceptions.ModificationNotAllowed("Cannot insert a path after storing the node")
-
-        if not os.path.isabs(src_abs):
-            raise ValueError("The source path in add_path must be absolute")
-        if os.path.isabs(dst_path):
-            raise ValueError("The destination path in add_path must be a" "filename without any subfolder")
-        self._get_folder_pathsubfolder().insert_path(src_abs, dst_path)
-
-    def get_abs_path(self, path=None, section=None):
-        """
-        Get the absolute path to the folder associated with the
-        Node in the AiiDA repository.
-
-        :param str path: the name of the subfolder inside the section. If None
-                         returns the abspath of the folder. Default = None.
-        :param section: the name of the subfolder ('path' by default).
-        :return: a string with the absolute path
-
-        For the moment works only for one kind of files, 'path' (internal files)
-        """
-        if path is None:
-            return self.folder.abspath
-        if section is None:
-            section = self._path_subfolder_name
-        # TODO: For the moment works only for one kind of files,
-        #      'path' (internal files)
-        if os.path.isabs(path):
-            raise ValueError("The path in get_abs_path must be relative")
-        return self.folder.get_subfolder(section, reset_limit=True).get_abs_path(path, check_existence=True)
+        return self._backend_entity.folder
 
     def store_all(self, with_transaction=True, use_cache=None):
         """
@@ -1197,96 +652,7 @@ class Node(entities.Entity):
         :parameter with_transaction: if False, no transaction is used. This is meant to be used ONLY if the outer
             calling function has already a transaction open!
         """
-        if self.is_stored:
-            raise exceptions.ModificationNotAllowed('Node<{}> is already stored'.format(self.id))
-
-        # For each node of a cached incoming link, check that all its incoming links are stored
-        for link_triple in self._incoming_cache:
-            try:
-                link_triple.node._check_are_parents_stored()
-            except exceptions.ModificationNotAllowed:
-                raise exceptions.ModificationNotAllowed(
-                    'source Node<{}> has unstored parents, cannot proceed (only direct parents can be unstored and '
-                    'will be stored by store_all, not grandparents or other ancestors'.format(link_triple.node.pk))
-
-        return self._db_store_all(with_transaction, use_cache=use_cache)
-
-    def _db_store_all(self, with_transaction=True, use_cache=None):
-        """
-        Store the node, together with all input links, if cached, and also the
-        linked nodes, if they were not stored yet.
-
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-
-        :param use_cache: Determines whether caching is used to find an equivalent node.
-        :type use_cache: bool
-        """
-        with self.backend.transaction():
-            self._store_input_nodes()
-            self.store()
-
-    def _store_input_nodes(self):
-        """
-        Find all input nodes, and store them, checking that they do not
-        have unstored inputs in turn.
-
-        :note: this function stores all nodes without transactions; always
-          call it from within a transaction!
-        """
-        if self.is_stored:
-            raise exceptions.ModificationNotAllowed(
-                'Node<{}> is already stored, but this method can only be called for unstored nodes'.format(self.pk))
-
-        for link_triple in self._incoming_cache:
-            if not link_triple.node.is_stored:
-                link_triple.node.store(with_transaction=False)
-
-    def _check_are_parents_stored(self):
-        """
-        Check if all parents are already stored, otherwise raise.
-
-        :raise ModificationNotAllowed: if one of the input nodes is not already stored.
-        """
-        for link_triple in self._incoming_cache:
-            if not link_triple.node.is_stored:
-                raise exceptions.ModificationNotAllowed(
-                    "Cannot store the incoming link triple {} because the source node is not stored. Either store it "
-                    "first, or call _store_input_links with `store_parents` set to True".format(link_triple.link_label))
-
-    def _store_cached_input_links(self, with_transaction=True):
-        """
-        Store all input links that are in the local cache, transferring them
-        to the DB.
-
-        :note: This can be called only if all parents are already stored.
-
-        :note: Links are stored only after the input nodes are stored. Moreover,
-            link storage is done in a transaction, and if one of the links
-            cannot be stored, an exception is raised and *all* links will remain
-            in the cache.
-
-        :note: This function can be called only after the node is stored.
-           After that, it can be called multiple times, and nothing will be
-           executed if no links are still in the cache.
-
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-        """
-        self._check_are_parents_stored()
-
-        # I have to store only those links where the source is already stored
-        for link_triple in self._incoming_cache:
-            self.backend_entity.add_link_from(*link_triple)
-
-        # If everything went smoothly, clear the entries from the cache.
-        # I do it here because I delete them all at once if no error
-        # occurred; otherwise, links will not be stored and I
-        # should not delete them from the cache (but then an exception
-        # would have been raised, and the following lines are not executed)
-        self._incoming_cache = list()
+        self._backend_entity.store_all(with_transaction=with_transaction, use_cache=use_cache)
 
     def store(self, with_transaction=True, use_cache=None):
         """
@@ -1305,198 +671,138 @@ class Node(entities.Entity):
           is meant to be used ONLY if the outer calling function has already
           a transaction open!
         """
-        # TODO: This needs to be generalized, allowing for flexible methods
-        # for storing data and its attributes.
+        self._backend_entity.store(with_transaction=with_transaction, use_cache=use_cache)
 
-        # As a first thing, I check if the data is storable
-        if not self._storable:
-            raise exceptions.StoringNotAllowed(self._unstorable_message)
+    # TODO: caching still to check
+    # def _store_from_cache(self, cache_node, with_transaction):
+    #     from aiida.orm.mixins import Sealable
+    #     assert self.type == cache_node.type
 
-        # Second thing: check if it's valid
-        self._validate()
+    #     self.label = cache_node.label
+    #     self.description = cache_node.description
 
-        if not self.is_stored:
+    #     for key, value in cache_node.iterattrs():
+    #         if key != Sealable.SEALED_KEY:
+    #             self._set_attr(key, value)
 
-            # Verify that parents are already stored. Raises if this is not the case.
-            self._check_are_parents_stored()
+    #     self.folder.replace_with_folder(cache_node.folder.abspath, move=False, overwrite=True)
 
-            # Get default for use_cache if it's not set explicitly.
-            if use_cache is None:
-                use_cache = get_use_cache(type(self))
-            # Retrieve the cached node.
-            same_node = self._get_same_node() if use_cache else None
-            if same_node is not None:
-                self._store_from_cache(same_node, with_transaction=with_transaction)
-                self._add_outputs_from_cache(same_node)
-            else:
-                # call implementation-dependent store method
-                self._db_store(with_transaction)
+    #     # Make sure the node doesn't have any RETURN links
+    #     if cache_node.get_outgoing(link_type=LinkType.RETURN).all():
+    #         raise ValueError('Cannot use cache from nodes with RETURN links.')
 
-            # Set up autogrouping used by verdi run
-            from aiida.orm.autogroup import current_autogroup, Autogroup, VERDIAUTOGROUP_TYPE
+    #     self.store(with_transaction=with_transaction, use_cache=False)
+    #     self.set_extra('_aiida_cached_from', cache_node.uuid)
 
-            if current_autogroup is not None:
-                if not isinstance(current_autogroup, Autogroup):
-                    raise exceptions.ValidationError("current_autogroup is not an AiiDA Autogroup")
+    # def _add_outputs_from_cache(self, cache_node):
+    #     # Add CREATE links
+    #     for entry in cache_node.get_outgoing(link_type=LinkType.CREATE):
+    #         new_node = entry.node.clone()
+    #         new_node.add_incoming(self, link_type=LinkType.CREATE, link_label=entry.link_label)
+    #         new_node.store()
 
-                if current_autogroup.is_to_be_grouped(self):
-                    group_name = current_autogroup.get_group_name()
-                    if group_name is not None:
-                        group = groups.Group.objects(self._backend).get_or_create(
-                            name=group_name, type_string=VERDIAUTOGROUP_TYPE)[0]
-                        group.add_nodes(self)
+    # def get_hash(self, ignore_errors=True, **kwargs):
+    #     """
+    #     Making a hash based on my attributes
+    #     """
+    #     from aiida.common.hashing import make_hash
+    #     try:
+    #         return make_hash(self._get_objects_to_hash(), **kwargs)
+    #     except Exception:  # pylint: disable=broad-except
+    #         if ignore_errors:
+    #             return None
+    #         else:
+    #             raise
 
-        return super(Node, self).store()
+    # def _get_objects_to_hash(self):
+    #     """
+    #     Return a list of objects which should be included in the hash.
+    #     """
+    #     computer = self.get_computer()
+    #     return [
+    #         importlib.import_module(self.__module__.split('.', 1)[0]).__version__, {
+    #             key: val
+    #             for key, val in self.get_attrs().items()
+    #             if (key not in self._hash_ignored_attributes and
+    #                 key not in getattr(self, '_updatable_attributes', tuple()))
+    #         }, self.folder, computer.uuid if computer is not None else None
+    #     ]
 
-    def _store_from_cache(self, cache_node, with_transaction):
-        from aiida.orm.mixins import Sealable
-        assert self.type == cache_node.type
+    # def rehash(self):
+    #     """
+    #     Re-generates the stored hash of the Node.
+    #     """
+    #     self.set_extra(_HASH_EXTRA_KEY, self.get_hash())
 
-        self.label = cache_node.label
-        self.description = cache_node.description
+    # def clear_hash(self):
+    #     """
+    #     Sets the stored hash of the Node to None.
+    #     """
+    #     self.set_extra(_HASH_EXTRA_KEY, None)
 
-        for key, value in cache_node.iterattrs():
-            if key != Sealable.SEALED_KEY:
-                self._set_attr(key, value)
+    # def get_cache_source(self):
+    #     """
+    #     Return the UUID of the node that was used in creating this node from the cache, or None if it was not cached
 
-        self.folder.replace_with_folder(cache_node.folder.abspath, move=False, overwrite=True)
+    #     :return: the UUID of the node from which this node was cached, or None if it was not created through the cache
+    #     """
+    #     return self.get_extra('_aiida_cached_from', None)
 
-        # Make sure the node doesn't have any RETURN links
-        if cache_node.get_outgoing(link_type=LinkType.RETURN).all():
-            raise ValueError('Cannot use cache from nodes with RETURN links.')
+    # @property
+    # def is_created_from_cache(self):
+    #     """
+    #     Return whether this node was created from a cached node.cached
 
-        self.store(with_transaction=with_transaction, use_cache=False)
-        self.set_extra('_aiida_cached_from', cache_node.uuid)
+    #     :return: boolean, True if the node was created by cloning a cached node, False otherwise
+    #     """
+    #     return self.get_cache_source() is not None
 
-    def _add_outputs_from_cache(self, cache_node):
-        # Add CREATE links
-        for entry in cache_node.get_outgoing(link_type=LinkType.CREATE):
-            new_node = entry.node.clone()
-            new_node.add_incoming(self, link_type=LinkType.CREATE, link_label=entry.link_label)
-            new_node.store()
+    # def _get_same_node(self):
+    #     """
+    #     Returns a stored node from which the current Node can be cached, meaning that the returned Node is a valid
+    #     cache, and its ``_aiida_hash`` attribute matches ``self.get_hash()``.
 
-    @abstractmethod
-    def _db_store(self, with_transaction=True):
-        """
-        Store a new node in the DB, also saving its repository directory
-        and attributes.
+    #     If there are multiple valid matches, the first one is returned. If no matches are found, ``None`` is returned.
 
-        After being called attributes cannot be
-        changed anymore! Instead, extras can be changed only AFTER calling
-        this store() function.
+    #     Note that after ``self`` is stored, this function can return ``self``.
+    #     """
+    #     try:
+    #         return next(self._iter_all_same_nodes())
+    #     except StopIteration:
+    #         return None
 
-        :note: After successful storage, those links that are in the cache, and
-            for which also the parent node is already stored, will be
-            automatically stored. The others will remain unstored.
+    # def get_all_same_nodes(self):
+    #     """
+    #     Return a list of stored nodes which match the type and hash of the current node. For the stored nodes, the
+    #     ``_aiida_hash`` extra is checked to determine the hash, while ``self.get_hash()`` is executed on the current
+    #     node.
 
-        :parameter with_transaction: if False, no transaction is used. This
-          is meant to be used ONLY if the outer calling function has already
-          a transaction open!
-        """
-        pass
+    #     Only nodes which are a valid cache are returned. If the current node is already stored, it can be included in
+    #     the returned list if ``self.get_hash()`` matches its ``_aiida_hash``.
+    #     """
+    #     return list(self._iter_all_same_nodes())
 
-    def get_hash(self, ignore_errors=True, **kwargs):
-        """
-        Making a hash based on my attributes
-        """
-        from aiida.common.hashing import make_hash
-        try:
-            return make_hash(self._get_objects_to_hash(), **kwargs)
-        except Exception:  # pylint: disable=broad-except
-            if ignore_errors:
-                return None
-            else:
-                raise
+    # def _iter_all_same_nodes(self):
+    #     """
+    #     Returns an iterator of all same nodes.
+    #     """
+    #     if not self._cacheable:
+    #         return iter(())
 
-    def _get_objects_to_hash(self):
-        """
-        Return a list of objects which should be included in the hash.
-        """
-        computer = self.get_computer()
-        return [
-            importlib.import_module(self.__module__.split('.', 1)[0]).__version__, {
-                key: val
-                for key, val in self.get_attrs().items()
-                if (key not in self._hash_ignored_attributes and
-                    key not in getattr(self, '_updatable_attributes', tuple()))
-            }, self.folder, computer.uuid if computer is not None else None
-        ]
+    #     hash_ = self.get_hash()
+    #     if not hash_:
+    #         return iter(())
 
-    def rehash(self):
-        """
-        Re-generates the stored hash of the Node.
-        """
-        self.set_extra(_HASH_EXTRA_KEY, self.get_hash())
+    #     builder = querybuilder.QueryBuilder()
+    #     builder.append(self.__class__, filters={'extras._aiida_hash': hash_}, project='*', subclassing=False)
+    #     same_nodes = (n[0] for n in builder.iterall())
+    #     return (n for n in same_nodes if n._is_valid_cache())  # pylint: disable=protected-access
 
-    def clear_hash(self):
-        """
-        Sets the stored hash of the Node to None.
-        """
-        self.set_extra(_HASH_EXTRA_KEY, None)
-
-    def get_cache_source(self):
-        """
-        Return the UUID of the node that was used in creating this node from the cache, or None if it was not cached
-
-        :return: the UUID of the node from which this node was cached, or None if it was not created through the cache
-        """
-        return self.get_extra('_aiida_cached_from', None)
-
-    @property
-    def is_created_from_cache(self):
-        """
-        Return whether this node was created from a cached node.cached
-
-        :return: boolean, True if the node was created by cloning a cached node, False otherwise
-        """
-        return self.get_cache_source() is not None
-
-    def _get_same_node(self):
-        """
-        Returns a stored node from which the current Node can be cached, meaning that the returned Node is a valid
-        cache, and its ``_aiida_hash`` attribute matches ``self.get_hash()``.
-
-        If there are multiple valid matches, the first one is returned. If no matches are found, ``None`` is returned.
-
-        Note that after ``self`` is stored, this function can return ``self``.
-        """
-        try:
-            return next(self._iter_all_same_nodes())
-        except StopIteration:
-            return None
-
-    def get_all_same_nodes(self):
-        """
-        Return a list of stored nodes which match the type and hash of the current node. For the stored nodes, the
-        ``_aiida_hash`` extra is checked to determine the hash, while ``self.get_hash()`` is executed on the current
-        node.
-
-        Only nodes which are a valid cache are returned. If the current node is already stored, it can be included in
-        the returned list if ``self.get_hash()`` matches its ``_aiida_hash``.
-        """
-        return list(self._iter_all_same_nodes())
-
-    def _iter_all_same_nodes(self):
-        """
-        Returns an iterator of all same nodes.
-        """
-        if not self._cacheable:
-            return iter(())
-
-        hash_ = self.get_hash()
-        if not hash_:
-            return iter(())
-
-        builder = querybuilder.QueryBuilder()
-        builder.append(self.__class__, filters={'extras._aiida_hash': hash_}, project='*', subclassing=False)
-        same_nodes = (n[0] for n in builder.iterall())
-        return (n for n in same_nodes if n._is_valid_cache())  # pylint: disable=protected-access
-
-    def _is_valid_cache(self):  # pylint: disable=no-self-use
-        """
-        Subclass hook to exclude certain Nodes (e.g. failed calculations) from being considered in the caching process.
-        """
-        return True
+    # def _is_valid_cache(self):  # pylint: disable=no-self-use
+    #     """
+    #     Subclass hook to exclude certain Nodes (e.g. failed calculations) from being considered in the caching process.
+    #     """
+    #     return True
 
     @property
     def out(self):
@@ -1525,12 +831,12 @@ class Node(entities.Entity):
         Property to understand if children are attached to the node
         :return: a boolean
         """
-        first_desc = querybuilder.QueryBuilder().append(
+        first_descendant = querybuilder.QueryBuilder().append(
             Node, filters={
                 'id': self.pk
             }, tag='self').append(
                 Node, with_ancestors='self', project='id').first()
-        return bool(first_desc)
+        return bool(first_descendant)
 
     @property
     def has_parents(self):
@@ -1545,6 +851,9 @@ class Node(entities.Entity):
                 Node, with_descendants='self', project='id').first()
         return bool(first_ancestor)
 
+    # TODO: decide if we want to keep it in the frontend and if we want
+    # to keep it with this name (or e.g. have simply .process, and
+    # maybe define it only in the ProcessNode subclass)
     def load_process_class(self):
         """
         For nodes that were ran by a Process, the process_type will be set. This can either be an entry point
@@ -1564,6 +873,38 @@ class Node(entities.Entity):
             process_class = getattr(module, class_name)
 
         return process_class
+
+    # region Links
+
+    def get_incoming(self, node_class=None, link_type=(), link_label_filter=None):
+        """
+        Return a list of link triples that are (directly) incoming into this node.
+
+        :param node_class: If specified, should be a class or tuple of classes, and it filters only
+            elements of that specific type (or a subclass of 'type')
+        :param link_type: If specified should be a string or tuple to get the inputs of this
+            link type, if None then returns all inputs of all link types.
+        :param link_label_filter: filters the incoming nodes by its link label.
+            Here wildcards (% and _) can be passed in link label filter as we are using "like" in QB.
+        """
+        return self._backend_entity.get_incoming(
+            node_class=node_class, link_type=link_type, link_label_filter=link_label_filter)
+
+    def get_outgoing(self, node_class=None, link_type=(), link_label_filter=None):
+        """
+        Return a list of link triples that are (directly) outgoing of this node.
+
+        :param node_class: If specified, should be a class or tuple of classes, and it filters only
+            elements of that specific type (or a subclass of 'type')
+        :param link_type: If specified should be a string or tuple to get the inputs of this
+            link type, if None then returns all outputs of all link types.
+        :param link_label_filter: filters the outgoing nodes by its link label.
+            Here wildcards (% and _) can be passed in link label filter as we are using "like" in QB.
+        """
+        return self._backend_entity.get_outgoing(
+            node_class=node_class, link_type=link_type, link_label_filter=link_label_filter)
+
+    # endregion
 
 
 #    @classmethod
@@ -1612,6 +953,8 @@ class Node(entities.Entity):
 #
 #        return self.id
 
+# region DeprecatedMethods
+
     @combomethod
     def querybuild(self_or_cls, **kwargs):  # pylint: disable=no-self-argument
         """
@@ -1633,7 +976,7 @@ class Node(entities.Entity):
         .. deprecated:: 1.0.0b1
             Use :meth:`aiida.orm.Node.objects.query` instead.
         """
-        warnings.warn("use aidia.orm.Node.objects.query instead", DeprecationWarning)
+        warnings.warn("use aiida.orm.Node.objects.query instead", DeprecationWarning)
 
         isclass = kwargs.pop('isclass')
         query = querybuilder.QueryBuilder()
@@ -1644,5 +987,198 @@ class Node(entities.Entity):
             filters.update({'id': self_or_cls.pk})
             query.append(self_or_cls.__class__, filters=filters, **kwargs)
         return query
+
+    @classmethod
+    def query(cls, *_args, **_kwargs):
+        """
+        Query for nodes of this type
+
+        .. deprecated:: 1.0.0b1
+            Use :meth:`aiida.orm.Node.objects.query` instead.
+        """
+        warnings.warn("use aidia.orm.Node.objects.query instead", DeprecationWarning)
+        raise NotImplementedError("This method is no longer available, use Node.objects.query()")
+
+    def get_inputs_dict(self, only_in_db=False, link_type=None):
+        """
+        Return a dictionary where the key is the label of the input link, and
+        the value is the input node.
+
+        .. deprecated:: 1.0.0b1
+
+        :param only_in_db: If true only get stored links, not cached
+        :param link_type: Only get inputs of this link type, if None then
+                returns all inputs of all link types.
+        :return: a dictionary {label:object}
+        """
+        warnings.warn('get_inputs_dict method is deprecated, use get_incoming instead', DeprecationWarning)
+
+        return dict(self.get_inputs(also_labels=True, only_in_db=only_in_db, link_type=link_type))
+
+    def get_outputs_dict(self, link_type=None):
+        """
+        Return a dictionary where the key is the label of the output link, and
+        the value is the input node.
+        As some Nodes (Datas in particular) can have more than one output with
+        the same label, all keys have the name of the link with appended the pk
+        of the node in output.
+        The key without pk appended corresponds to the oldest node.
+
+        .. deprecated:: 1.0.0b1
+
+        :return: a dictionary {linkname:object}
+        """
+        warnings.warn('get_outputs_dict method is deprecated, use get_outgoing instead', DeprecationWarning)
+
+        if link_type is not None and not isinstance(link_type, LinkType):
+            raise TypeError("link_type should be a LinkType object")
+
+        all_outputs = self.get_outputs(also_labels=True, link_type=link_type)
+
+        all_linknames = [i[0] for i in all_outputs]
+        linknames_set = list(set(all_linknames))
+
+        # prepare a new output list
+        new_outputs = {}
+        # first add the defaults
+        for irreducible_linkname in linknames_set:
+            this_elements = [i[1] for i in all_outputs if i[0] == irreducible_linkname]
+            # select the oldest element
+            last_element = sorted(this_elements, key=lambda x: x.ctime)[0]
+            # for this one add the default value
+            new_outputs[irreducible_linkname] = last_element
+
+            # now for everyone append the string with the pk
+            for i in this_elements:
+                new_outputs[irreducible_linkname + "_{}".format(i.pk)] = i
+
+        return new_outputs
+
+    def get_inputs(self, node_type=None, also_labels=False, only_in_db=False, link_type=None):
+        """
+        Return a list of nodes that enter (directly) in this node
+
+        :param node_type: If specified, should be a class, and it filters only
+            elements of that specific type (or a subclass of 'type')
+        :param also_labels: If False (default) only return a list of input nodes.
+            If True, return a list of tuples, where each tuple has the
+            following format: ('label', Node), with 'label' the link label,
+            and Node a Node instance or subclass
+        :param only_in_db: Return only the inputs that are in the database,
+            ignoring those that are in the local cache. Otherwise, return
+            all links.
+        :param link_type: Only get inputs of this link type, if None then
+            returns all inputs of all link types.
+
+        .. deprecated:: 1.0.0b1
+        """
+        warnings.warn('get_inputs method is deprecated, use get_incoming instead', DeprecationWarning)
+
+        if link_type is not None and not isinstance(link_type, LinkType):
+            raise TypeError('link_type should be a LinkType object')
+
+        inputs_list = self._backend_entity._get_db_input_links(link_type=link_type)
+
+        if not only_in_db:
+            # Needed for the check
+            input_list_keys = [i[0] for i in inputs_list]
+
+            for link_triple in self._incoming_cache:
+                if link_triple.link_label in input_list_keys:
+                    raise exceptions.InternalError(
+                        "There exist a link with the same name '{}' both in the DB and in the internal "
+                        "cache for node pk= {}!".format(link_triple.link_label, self.pk))
+
+                if link_type is None or link_triple.link_type is link_type:
+                    inputs_list.append((link_triple.link_label, link_triple.node))
+
+        if node_type is None:
+            filtered_list = inputs_list
+        else:
+            filtered_list = [i for i in inputs_list if isinstance(i[1], node_type)]
+
+        if also_labels:
+            return list(filtered_list)
+
+        return [i[1] for i in filtered_list]
+
+    def get_outputs(self, node_type=None, also_labels=False, link_type=None):
+        """
+        Return a list of nodes that exit (directly) from this node
+
+        :param node_type: if specified, should be a class, and it filters only
+            elements of that specific node_type (or a subclass of 'node_type')
+        :param also_labels: if False (default) only return a list of input nodes.
+            If True, return a list of tuples, where each tuple has the
+            following format: ('label', Node), with 'label' the link label,
+            and Node a Node instance or subclass
+        :param link_type: Only return outputs connected by links of this type.
+
+        .. deprecated:: 1.0.0b1
+        """
+        warnings.warn('get_outputs method is deprecated, use get_outgoing instead', DeprecationWarning)
+
+        if link_type is not None and not isinstance(link_type, LinkType):
+            raise TypeError('link_type should be a LinkType object')
+
+        outputs_list = [
+            convert.get_orm_entity(entity) for entity in self.backend_entity._get_db_output_links(link_type=link_type)
+        ]
+
+        if node_type is None:
+            filtered_list = outputs_list
+        else:
+            filtered_list = (i for i in outputs_list if isinstance(i[1], node_type))
+
+        if also_labels:
+            return list(filtered_list)
+
+        return [i[1] for i in filtered_list]
+
+    def iterattrs(self):
+        """
+        Iterator over the Node attributes, returning tuples (key, value)
+
+        .. deprecated:: 1.0.0b1
+        """
+        warnings.warn('iterattrs has been deprecated, use attrs_items instead', DeprecationWarning)
+
+        for item in self._backend_entity.attrs_items:
+            yield item
+
+    def attrs(self):
+        """
+        Returns the keys of the attributes as a generator.
+
+        .. deprecated:: 1.0.0b1
+        :return: a generator of a strings
+        """
+        warnings.warn('iterattrs has been deprecated, use attrs_keys instead', DeprecationWarning)
+
+        for item in self._backend_entity.attrs_keys:
+            yield item
+
+    def extras(self):
+        """
+        Get the keys of the extras.
+
+        .. deprecated:: 1.0.0b1
+        :return: a list of strings
+        """
+        warnings.warn('extras has been deprecated, use extras_keys instead', DeprecationWarning)
+
+        for item in self._backend_entity.extras_items:
+            yield item
+
+    def iterextras(self):
+        """
+        Iterator over the extras, returning tuples (key, value)
+
+        .. deprecated:: 1.0.0b1
+        """
+        warnings.warn('iterextras has been deprecated, use extras_items instead', DeprecationWarning)
+
+        for item in self._backend_entity.extras_keys:
+            yield item
 
     # endregion
